# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gmail/send:
    post:
      operationId: sendGmailDemoEmail
      tags: [gmail]
      summary: Send an approved synthetic demo email
      description: Sends a reviewed synthetic message to the connected Gmail account itself. The client must require explicit approval before calling this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GmailDemoEmailInput"
      responses:
        "200":
          description: Demo email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailSendResult"
        "400":
          description: Invalid or unsafe message
        "503":
          description: Gmail connection unavailable
...
  /calendar/follow-ups:
    post:
      operationId: createCalendarFollowUp
      tags: [calendar]
      summary: Create an approved synthetic follow-up event
      description: Creates a no-attendee event from an already-approved persisted SchoolOps action, then marks that action completed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarFollowUpApproval"
      responses:
        "200":
          description: Follow-up event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarFollowUpResult"
        "400":
          description: Invalid or unsafe follow-up request
        "404":
          description: Approved action not found
        "409":
          description: Action has not been approved or was already completed
        "503":
          description: Google Calendar connection unavailable
...
components:
  schemas:
...
    GmailDemoEmailInput:
      type: object
      properties:
        recipient:
          type: string
          const: connected-test-account
        subject:
          type: string
          minLength: 1
          maxLength: 160
        body:
          type: string
          minLength: 1
          maxLength: 5000
        approved:
          type: boolean
          const: true
        syntheticDataOnly:
          type: boolean
          const: true
      required: [recipient, subject, body, approved, syntheticDataOnly]
    GmailSendResult:
      type: object
      properties:
        sent:
          type: boolean
        messageId:
          type: string
        recipientLabel:
          type: string
      required: [sent, messageId, recipientLabel]
...
    CalendarFollowUpApproval:
      type: object
      properties:
        actionId:
          type: integer
        approved:
          type: boolean
          const: true
        syntheticDataOnly:
          type: boolean
          const: true
      required: [actionId, approved, syntheticDataOnly]
    CalendarFollowUpResult:
      type: object
      properties:
        created:
          type: boolean
        eventId:
          type: string
        calendarLabel:
          type: string
        actionId:
          type: integer
        completedAt:
          type: string
          format: date-time
      required: [created, eventId, calendarLabel, actionId, completedAt]
...